changes to @kubernetes for sandbox (#1079)
